requirements/base.txt
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.24.0
billiard==3.3.0.23
boto==2.42.0
botocore==1.13.32
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
colorama==0.3.9
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
cryptography==2.3.1
defusedxml==0.5.0
dj-database-url==0.5.0
Django==1.11.1
django-allauth==0.36.0
django-bootstrap3==9.0.0
django-classy-tags==0.8.0
django-cms==3.5.2
django-cors-headers==2.4.0
django-crispy-forms==1.7.2
django-crontab==0.7.1
django-extensions==2.0.7
django-filer==1.3.2
django-filter==1.1.0
django-forms-bootstrap==3.0.1
django-formtools==2.1
django-guardian==1.5.0
django-last-seen==0.3
django-meta==1.4.1
django-mptt==0.8.7
django-paypal==0.3.6
django-polymorphic==2.0.2
django-rest-auth==0.9.3
django-sekizai==0.10.0
django-sql-explorer==1.1.1
django-storages==1.5.1
django-treebeard==4.3
django-tz-detect==0.2.7
django-webpack-loader==0.5.0
django-werkzeug-debugger-runserver==0.3.1
djangocms-accordion==1.0.4
djangocms-admin-style==1.2.8
djangocms-attributes-field==0.3.0
djangocms-bootstrap4==1.0.0
djangocms-column==1.7.0
djangocms-file==2.0.2
djangocms-googlemap==1.1.1
djangocms-helper==0.9.9
djangocms-history==0.5.3
djangocms-icon==1.0.0
djangocms-inherit==0.2.2
djangocms-installer==1.0.2
djangocms-link==2.1.2
djangocms-page-meta==0.8.3
djangocms-picture==2.0.6
djangocms-snippet==2.0.0
djangocms-style==2.0.2
djangocms-table==1.2
djangocms-teaser==0.2.0
djangocms-text-ckeditor==3.6.0
djangocms-video==2.0.4
djangorestframework==3.8.2
docopt==0.6.2
docutils==0.14
docxtpl==0.5.17
easy-thumbnails==2.5
ecdsa==0.13
future==0.16.0
gunicorn==20.0.4
html5lib==0.9999999
idna==2.6
itypes==1.1.0
jaraco.classes==2.0
jaraco.collections==2.0
jaraco.functools==2.0
jaraco.text==3.0
Jinja2==2.10
jmespath==0.9.4
jpush==3.2.8
kombu==3.0.37
lxml==4.4.2
Markdown==2.6.6
MarkupSafe==1.1.1
more-itertools==6.0.0
oauthlib==2.0.6
olefile==0.45.1
openpack==1.1.3
opentok==2.6.0
optionaldict==0.1.1
pip==19.3.1
pluggy==0.6.0
py==1.5.3
pyasn1==0.4.3
pycparser==2.18
pycryptodome==3.7.3
pyflakes==2.0.0
PyMySQL==0.9.1
pyOpenSSL==18.0.0
python-dateutil==2.7.3
python-docx==0.8.6
python-jose==3.0.0
python-pptx==0.6.17
python3-openid==3.1.0
pytz==2018.4
PyYAML==3.13
qrcode==6.0
raven==6.10.0
redis==2.10.6
reportlab==3.4.0
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
s3transfer==0.2.0
setuptools==42.0.2
six==1.11.0
South==1.0.2
sqlparse==0.2.4
stripe==2.38.0
tox==3.0.0
tzlocal==1.5.1
unicodecsv==0.14.1
Unidecode==0.4.21
uritemplate==3.0.0
urllib3==1.22
vine==1.1.4
virtualenv==16.0.0
Wand==0.5.1
wechatpy==1.7.0
Werkzeug==0.14.1
wheel==0.33.6
XlsxWriter==1.1.5
xmltodict==0.11.0
openpyxl==3.0.4
gevent==20.6.2